+++ ReleaseNotes.md
@@ -4,6 +4,10 @@ This page highlights some changes in the SDK.
 
 Not all changes will be listed, but you can always [compare by version tags](https://github.com/AquaticInformatics/aquarius-sdk-net/compare/v17.2.21...v17.2.25) to see the full source code difference.
 
+### 20.2.0
+- Updated the service models for the AQUARIUS Time-Series 2020.2 release.
+- Updated the service models for the AQUARIUS Samples 2020.05 release.
+
 ### 20.1.0
 - Updated the service models for the AQUARIUS Time-Series 2020.1 release.